Déplace le code dans un composant
